% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCIMIS.R
\name{findCIMIS}
\alias{findCIMIS}
\title{Find CIMIS Stations}
\usage{
findCIMIS(AOI = NULL, ids = FALSE)
}
\arguments{
\item{AOI}{A Spatial* or simple features geometry, can be piped from \link[AOI]{getAOI}}

\item{ids}{If TRUE, a vector of CIMIS station IDs are added to retuned list (default = \code{FALSE})}
}
\value{
a list() of minimum length 2: AOI and nwis
}
\description{
\code{findCIMIS} returns a \code{SpatialPointsDataFrame}
of all California Department of Water Resources CIMIS stations for an Area of Interest. This dataset is accessed through the CIMIS FTP 
and contains the following attributes:
\itemize{
\item 'Station Number':      \code{character}  Unique Station ID
\item 'DWR Regional Office': \code{character}  Department of Water Resources Regional Office
\item 'Name':                \code{character}  Station Name
\item 'County':              \code{character}  Station County
\item 'Latitude':            \code{numeric}    Station latitude, decimil degrees
\item 'Longitude':           \code{numeric}    Station longitude, decimil degrees
\item 'Elevation':           \code{numeric}    Station Elevation (meters)
\item 'Status':              \code{logical}    Is the station active?
\item 'Connect':             \code{Date}       Date of first observation
\item 'Disconnect':          \code{Date}       Date of last observation
} \cr
}
\examples{
\dontrun{
sb.cimis = getAOI(state = "CA", county = 'Santa Barbara') \%>\% findCIMIS()
}
}
\author{
Mike Johnson
}
