% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNearestCIMIS.R
\name{findNearestCIMIS}
\alias{findNearestCIMIS}
\title{Find Neareast CIMIS station(s)}
\usage{
findNearestCIMIS(point = NULL, n = 5, ids = FALSE, bb = FALSE)
}
\arguments{
\item{point}{a point described by lat/long, can be piped from \link[AOI]{geocode}}

\item{n}{the number of stations to find (default = 5)}

\item{ids}{If TRUE, a vector of station IDs is added to retuned list (default = \code{FALSE})}

\item{bb}{If TRUE, the geometry of the minimum bounding area of the features is added to returned list  (default = \code{FALSE})}
}
\value{
a list() of minimum length 2: AOI and cimis
}
\description{
\code{findNearestCIMIS} returns a \code{SpatialPointsDataFrame} of the 'n' number of CIMIS stations closest to a declared point.
}
\examples{
\dontrun{
pt = geocode("UCSB") \%>\% findNearestCIMIS(n = 5)
}
}
\seealso{
findCIMIS
}
\author{
Mike Johnson
}
